+++ lib/src/dart_resources.dart
@@ -1,4 +1,16 @@
-part of discovery_api_client_generator;
+// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library discoveryapis_generator.dart_resources;
+
+import 'generated_googleapis/discovery/v1.dart';
+import 'dart_comments.dart';
+import 'dart_schemas.dart';
+import 'dart_api_library.dart';
+import 'namer.dart';
+import 'uri_template.dart';
+import 'utils.dart';
 
 const RESERVED_METHOD_PARAMETER_NAMES =
     const ['uploadMedia', 'uploadOptions', 'downloadOptions', 'callOptions'];