+++ lib/src/dart_schemas.dart
@@ -1,5 +1,14 @@
-part of discovery_api_client_generator;
+// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
 
+library discoveryapis_generator.dart_schemas;
+
+import 'generated_googleapis/discovery/v1.dart';
+import 'dart_comments.dart';
+import 'dart_api_library.dart';
+import 'namer.dart';
+import 'utils.dart';
 
 /**
  * Class for keeping all named schemas. This is used for