+++ lib/src/namer.dart
@@ -1,4 +1,10 @@
-part of discovery_api_client_generator;
+// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+library discoveryapis_generator.namer;
+
+import 'utils.dart';
 
 /**
  * Represents an identifier that can be given a name.