+++ test/src/namer_test.dart
@@ -1,6 +1,9 @@
+// Copyright (c) 2015, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:discoveryapis_generator/src/namer.dart';
 import 'package:unittest/unittest.dart';
-import 'package:discovery_api_client_generator/generator.dart'
-    show ApiLibraryNamer, Scope, Identifier, IdentifierNamer;
 
 main() {
   group('namer', () {